refactor config system so that all the global variables are defined in manim/config/__init__.py. All other files inside manim/config/ just contain utils that are used by either manim/config/__init__.py or directly by manim/__main__.py
