+++ tests/test_cli_flags.py
@@ -1,45 +1,45 @@
 from manim import constants
-from manim.config.config_utils import _determine_quality, _parse_cli
+from manim.config.main_utils import _determine_quality, parse_args
 
 
 def test_quality_flags():
     # Assert that quality is the default when not specifying it
-    parsed = _parse_cli([], False)
+    parsed = parse_args('manim dummy_filename'.split())
 
     assert parsed.quality == constants.DEFAULT_QUALITY_SHORT
     assert _determine_quality(parsed) == constants.DEFAULT_QUALITY
 
     for quality in constants.QUALITIES.keys():
         # Assert that quality is properly set when using -q*
-        arguments = f"-q{constants.QUALITIES[quality]}".split()
-        parsed = _parse_cli(arguments, False)
+        arguments = f"manim -q{constants.QUALITIES[quality]} dummy_filename".split()
+        parsed = parse_args(arguments)
 
         assert parsed.quality == constants.QUALITIES[quality]
         assert quality == _determine_quality(parsed)
 
         # Assert that quality is properly set when using -q *
-        arguments = f"-q {constants.QUALITIES[quality]}".split()
-        parsed = _parse_cli(arguments, False)
+        arguments = f"manim -q {constants.QUALITIES[quality]} dummy_filename".split()
+        parsed = parse_args(arguments)
 
         assert parsed.quality == constants.QUALITIES[quality]
         assert quality == _determine_quality(parsed)
 
         # Assert that quality is properly set when using --quality *
-        arguments = f"--quality {constants.QUALITIES[quality]}".split()
-        parsed = _parse_cli(arguments, False)
+        arguments = f"manim --quality {constants.QUALITIES[quality]} dummy_filename".split()
+        parsed = parse_args(arguments)
 
         assert parsed.quality == constants.QUALITIES[quality]
         assert quality == _determine_quality(parsed)
 
         # Assert that quality is properly set when using -*_quality
-        arguments = f"--{quality}".split()
-        parsed = _parse_cli(arguments, False)
+        arguments = f"manim --{quality} dummy_filename".split()
+        parsed = parse_args(arguments)
 
         assert getattr(parsed, quality)
         assert quality == _determine_quality(parsed)
 
         # Assert that *_quality is False when not specifying it
-        parsed = _parse_cli([], False)
+        parsed = parse_args('manim dummy_filename'.split())
 
         assert not getattr(parsed, quality)
         assert _determine_quality(parsed) == constants.DEFAULT_QUALITY
